Added tab complete for fill inv
